Create HBV_quasispecies_sfs.py
Place all vcf and coverage files for analysis in one folder. Title the files as [ID_genotype.***.vcf] and [ID_genotype.***.coverage]. The output will simply look like
41      0       0       0       704  #non-variable site
42      0       0       694.823786      9.176214 #variable site
43      0       0       0       704
44      0       0       0       704

Col1: reference position
Col2: counts of A in population
Col3: counts of C in population
Col4: counts of G in population
Col5: counts of T in population

Where every individual sample is a percentage of 1. (A viral population with a SNP at position 869, ref A, 99% C, will be recorded as 0.99 C and 0.01 A.) Values across all individuals in analysis are added.